backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
